+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- [**BC**] Change `Command` type
+    - Remove `Synchronous/Asynchronous` variants
+- [**BC**] Rename `CommandHandleResult` cases
+    - `AsynchronousCommandStarted` -> `CommandStarted`
+    - `AsynchronousCommandNotStarted` -> `CommandNotStarted`
+    - `SynchronousCommandResponse` -> `CommandResponse`
 
 ## 5.8.0 - 2021-09-29
 - Add `DataItemDto.serializeResult` function